role commit

+++ src/app/master/shift/shift.component.ts
@@ -1,15 +1,14 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild ,ChangeDetectorRef  } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTable, MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Shift } from 'src/app/models/shift.model';
-import { DialogService } from 'src/app/services/dialog.service';
 import { ShiftApiService } from 'src/app/services/shift-api.service';
 import { AlertService } from 'src/app/shared/alert.service';
 import { AddShiftComponent } from './add-shift/add-shift.component';
-
+import { DialogsService } from 'src/app/services/dialogs.service';
 
 
 const ELEMENT_DATA: Shift[] = [];
@@ -23,8 +22,6 @@ export class ShiftComponent implements OnInit {
 
   dataShift: Shift[] = [];
 
-  private refresh
-
   shiftDataSource;
 
   displayedColumns: string[] = ['sno', 'shiftName', 'startTime', 'endTime', 'actions'];
@@ -33,13 +30,15 @@ export class ShiftComponent implements OnInit {
   @ViewChild(MatSort) sort: MatSort;
   @ViewChild(MatTable, { static: true }) table: MatTable<any>;
 
+
   constructor(
     private shiftApi: ShiftApiService,
     private router: Router,
     private activatedRoute: ActivatedRoute,
     public dialog: MatDialog,
-    private dialogService: DialogService,
-    private alert: AlertService
+    private alert: AlertService,
+    private changeDetectorRef:ChangeDetectorRef,
+    private dialogsService:DialogsService
   ) {
 
   }
@@ -55,17 +54,13 @@ export class ShiftComponent implements OnInit {
 
   }
 
-
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.shiftDataSource.filter = filterValue.trim().toLowerCase();
   }
 
-
   onClickAdd() {
-
     this.dialog.open(AddShiftComponent);
-
   }
 
 
@@ -73,13 +68,20 @@ export class ShiftComponent implements OnInit {
     this.dialog.open(AddShiftComponent, { data: { shift } });
   }
 
-
   onClickDelete(id: string) {
 
-    this.shiftApi.deleteShift(id).subscribe(res => {
-      this.dataShift = this.dataShift.filter(item => item._id !== id);
-      this.alert.showError('Data Deleted Suceessfully...!', 'Shift');
-    });
-  }
+  
+  this.dialogsService.openConfirmDialog('Are you sure to delete this record ?')
+  .afterClosed().subscribe(res => {
+    if(res){
+      this.shiftApi.deleteShift(id).subscribe(res => {
+        this.dataShift = this.dataShift.filter(item => item._id !== id);
+        this.ngOnInit();
+        this.alert.showError('Data Deleted Suceessfully...!', 'Shift');
+      })
+    }
+  });
+
+}
 
 }
